feat: Add confirmation on clear

+++ Lessons/lesson-21-18-06-2021/to-do-list/js/todolist.js
@@ -151,8 +151,10 @@ list.addEventListener("click", function (event) {
 
 //Clear the local storage
 clear.addEventListener("click", function () {
-  localStorage.clear();
-  location.reload();
+  if (window.confirm("Clear all tasks?")) {
+    localStorage.clear();
+    location.reload();
+  }
 });
 
 // Filter
